Testing ways to make query output usable.
printr outputs a string and not an associative array, this means that I can't access any of the variables found in the query.
With this in mind I'm probably going to have to filter out all of the non-numeric characters and put those into variables manually, then put those into the cookies.