Completed database integration and data transfer from page to page
We can now get through from dndbuilder to class to dice. All values chosen from one page move on to the next, and so on.
I had to make the data transfer as an array, which makes the code much more readable. Happy side effect.
Also fixed the dice reroll issue. It was redirecting to multidice.php which doesn't exist in our files.