Dice now displays everything but the stat modifiers. Also connected to our stylesheet.
I'll get the calculation and the table in for that. It'll be really easy. 
The not so easy part will be the CSS. We need to hammer that down. I'll do what I can but I'm not sure what all Tom's done. 
I also tried splitting weapons because that could get long, but it wouldn't go on to a new line when I used \n. I'm not sure how to fix that but I left my code in the dice.blade.php section.